// Jobs.Frontend/src/app/core/api/model/jobPostingResponsePaginatedListResult.ts
/**
 * Jobs API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JobPostingResponse } from './jobPostingResponse';


/**
 * Represents a paginated result set.
 */
export interface JobPostingResponsePaginatedListResult { 
    items?: Array<JobPostingResponse> | null;
    totalCount?: number;
    pageNumber?: number;
    pageSize?: number;
    totalPages?: number;
}